Convert sample type select rendering code to module and import into geneVariant handler
