+++ client/dom/index.ts
@@ -36,6 +36,7 @@ export * from './renderAtRisk'
 export * from './renderPvalueTable'
 export * from './resultsTable'
 export * from './types/resultsTable'
+export * from './sampleTypeSelect'
 export * from './sandbox'
 export * from './sayerror'
 export * from './search'
